6.1.14 Выносит общую логику обновления в smat компонент
